WIP: Work on matching interfaces nearly done [skip ci]
