+++ openghg/client/__init__.py
@@ -1,7 +1,6 @@
-from ._user import create_user
 from ._jobrunner import JobRunner
-from ._process import Process
-from ._retrieve import Retrieve
-from ._search import Search
-from ._rank_sources import RankSources
-from ._test_service import TestService
+from ._process import process_files
+from ._rank_sources import rank_sources
+from ._search import search
+from ._test_service import call_test_service
+from ._user import create_user
