+++ openghg/client/_rank_sources.py
@@ -1,19 +1,22 @@
-from typing import Dict, List, Optional, Union
-from openghg.util import valid_site, create_daterange_str, InvalidSiteError
+from dataclasses import dataclass
+from typing import Dict, List, Union
 
-from Acquire.Client import Wallet
-
-__all__ = ["RankSources"]
+from openghg.store import ObsSurface
+from openghg.store.base import Datasource
+from openghg.util import InvalidSiteError, create_daterange_str, running_in_cloud, valid_site
 
 
 class RankSources:
-    def __init__(self, service_url: Optional[str] = None):
-        wallet = Wallet()
+    def __init__(self, cloud: bool = False) -> None:
+        self._cloud = cloud
 
-        if service_url is None:
+        if cloud:
+            from Acquire.Client import Wallet
+
+            wallet = Wallet()
             service_url = "https://fn.openghg.org/t"
 
-        self._service = wallet.get_service(service_url=f"{service_url}/openghg")
+            self._service = wallet.get_service(service_url=f"{service_url}/openghg")
 
     def get_sources(self, site: str, species: str) -> Dict:
         """Get the datasources for this site and species to allow a ranking to be set
@@ -24,25 +27,65 @@ class RankSources:
         Returns:
             dict: Dictionary of datasource metadata
         """
+        if self._cloud:
+            return self._get_sources_cloud(site=site, species=species)
+        else:
+            return self._get_sources_local(site=site, species=species)
+
+    def _get_sources_cloud(self, site: str, species: str) -> Dict:
         if not valid_site(site):
             raise InvalidSiteError(f"{site} is not a valid site code")
 
         args = {"site": site, "species": species}
 
-        response = self._service.call_function(function="rank.get_sources", args=args)
+        self.site = site
+        self.species = species
+
+        response: Dict = self._service.call_function(function="rank.get_sources", args=args)
 
         if not response:
             raise ValueError(f"No sources found for {species} at {site}")
 
         self._user_info: Dict = response["user_info"]
         self._key_lookup: Dict = response["key_lookup"]
-
-        self._lookup_data = {"site": site, "species": species}
         self._needs_update = False
 
         return self._user_info
 
-    def get_specific_source(self, key: str) -> str:
+    def _get_sources_local(self, site: str, species: str) -> Dict:
+        if not valid_site(site):
+            raise InvalidSiteError(f"{site} is not a valid site code")
+
+        # Save these
+        self.site = site
+        self.species = species
+
+        obs = ObsSurface.load()
+        datasource_uuids = obs.datasources()
+        rank_table = obs.rank_data()
+
+        # Shallow load the Datasources (only get their JSON metadata)
+        datasources = (Datasource.load(uuid=uuid, shallow=True) for uuid in datasource_uuids)
+
+        matching_sources = [d for d in datasources if d.search_metadata(site=site, species=species)]
+
+        if not matching_sources:
+            return {}
+
+        self._user_info = {
+            d.inlet(): {
+                "rank_data": rank_table.get(d.uuid(), "NA"),
+                "data_range": d.daterange_str(),
+            }
+            for d in matching_sources
+        }
+
+        self._key_lookup = {d.inlet(): d.uuid() for d in matching_sources}
+        self._needs_update = False
+
+        return self._user_info
+
+    def get_specific_source(self, inlet: str) -> Dict:
         """Return the ranking data of a specific key
 
         Args:
@@ -51,41 +94,59 @@ class RankSources:
             dict: Dictionary of ranking data
         """
         if self._needs_update:
-            site = self._lookup_data["site"]
-            species = self._lookup_data["species"]
-            _ = self.get_sources(site=site, species=species)
+            self.get_sources(site=self.site, species=self.species)
 
-        rank_data: str = self._user_info[key]["rank_data"]
-
+        rank_data: Dict[str, Union[str, Dict]] = self._user_info[inlet]["rank_data"]
         return rank_data
 
     def set_rank(
         self,
-        key: str,
+        inlet: str,
         rank: Union[int, str],
-        start_date: Optional[str] = None,
-        end_date: Optional[str] = None,
+        start_date: str,
+        end_date: str,
         overwrite: bool = False,
-        dateranges: Optional[Union[List, str]] = None,
     ) -> None:
         """Set the rank data for the
 
         Args:
-            key: Key of ranking data from the original dict
-            return by get_sources.
+            inlet: Inlet to set ranking data
             rank: Number between 1 and 9
             start_date: Start date
             end_date: End date
+            overwrite: If True overwrite current ranking data
         Returns:
             None
         """
-        if all((start_date, end_date, dateranges)):
-            raise ValueError("Either a start and end date must be passed or a list of dateranges")
+        if self._cloud:
+            return self._set_rank_cloud(
+                inlet=inlet,
+                rank=rank,
+                start_date=start_date,
+                end_date=end_date,
+                overwrite=overwrite,
+            )
+        else:
+            return self._set_rank_local(
+                inlet=inlet,
+                rank=rank,
+                start_date=start_date,
+                end_date=end_date,
+                overwrite=overwrite,
+            )
 
-        uuid = self._key_lookup[key]
+    def _set_rank_cloud(
+        self,
+        inlet: str,
+        rank: Union[int, str],
+        start_date: str,
+        end_date: str,
+        overwrite: bool = False,
+    ) -> None:
+        inlet = inlet.lower()
+        uuid = self._key_lookup[inlet]
 
-        if dateranges is None:
-            dateranges = create_daterange_str(start=start_date, end=end_date)
+        dateranges = create_daterange_str(start=start_date, end=end_date)
 
         args: Dict[str, Union[str, int, List]] = {}
         args["rank"] = rank
@@ -96,7 +157,26 @@ class RankSources:
         self._service.call_function(function="rank.set_rank", args=args)
         self._needs_update = True
 
-    def clear_rank(self, key: str) -> None:
+    def _set_rank_local(
+        self,
+        inlet: str,
+        rank: Union[int, str],
+        start_date: str,
+        end_date: str,
+        overwrite: bool = False,
+    ) -> None:
+        obs = ObsSurface.load()
+
+        inlet = inlet.lower()
+        uuid = self._key_lookup[inlet]
+
+        daterange = create_daterange_str(start=start_date, end=end_date)
+
+        obs.set_rank(uuid=uuid, rank=rank, date_range=daterange, overwrite=overwrite)
+
+        self._needs_update = True
+
+    def clear_rank(self, inlet: str) -> None:
         """Clear the ranking data for a Datasource
 
         Args:
@@ -104,8 +184,102 @@ class RankSources:
         Returns:
             None
         """
-        uuid = self._key_lookup[key]
+        if self._cloud:
+            return self._clear_rank_cloud(inlet=inlet)
+        else:
+            return self._clear_rank_local(inlet=inlet)
 
+    def _clear_rank_local(self, inlet: str) -> None:
+        """Clear the ranking data for a Datasource
+
+        Args:
+            key: Key for specific source
+        Returns:
+            None
+        """
+        obs = ObsSurface.load()
+        inlet = inlet.lower()
+        uuid = self._key_lookup[inlet]
+        obs.clear_rank(uuid=uuid)
+        self._needs_update = True
+
+    def _clear_rank_cloud(self, inlet: str) -> None:
+        """Clear the ranking data for a Datasource
+
+        Args:
+            key: Key for specific source
+        Returns:
+            None
+        """
+        uuid = self._key_lookup[inlet]
         args = {"uuid": uuid}
-
         self._service.call_function(function="rank.clear_rank", args=args)
+        self._needs_update = True
+
+    # def visualise_rankings(self) -> Network:
+    #     """ Creates a small network graph of ranked data with each rank given a colour
+
+    #         Note that this function should only be run from a Jupyter Notebook
+
+    #         Args:
+    #             rank_data (dict): Dictionary of the form given by RankSources.get_sources()
+    #         Returns:
+    #             pyvis.network.Network: Network graph
+    #     """
+    #     header_text = "OpenGHG ranked data"
+    #     net = Network("800px", "100%", notebook=True, heading=header_text)
+    #     # Set the physics layout of the network
+    #     net.force_atlas_2based()
+
+    #     rank_data = self._key_lookup
+
+    #     a_key = list(rank_data.keys())[0]
+    #     site = rank_data[a_key]["metadata"]["site"].upper()
+
+    #     norm = matplotlib.colors.Normalize(vmin=0, vmax=10, clip=True)
+    #     mapper = cm.ScalarMappable(norm=norm, cmap=cm.tab10)
+
+    #     def colour_mapper(x):
+    #         return matplotlib.colors.to_hex(mapper.to_rgba(int(x)))
+
+    #     net.add_node(site, label=site, color="brown", value=5000)
+
+    #     for key, data in rank_data.items():
+    #         rank = data["rank"]
+    #         site_name = data["metadata"]["site"].upper()
+    #         data_range = data["data_range"]
+
+    #         # HTML to show when the mouse is hovered over a node
+    #         title = "</br>".join([f"<b>Rank:</b> {str(rank)}", f"<b>Site:</b> {site_name}", f"<b>Data range:</b> {data_range}"])
+
+    #         if rank == 0:
+    #             colour = colour_mapper(rank)
+    #         else:
+    #             # For now just use the highest rank for the color
+    #             highest_rank = sorted(list(rank.keys()))[-1]
+    #             colour = colour_mapper(highest_rank)
+
+    #         split_key = key.split("_")
+    #         label = " ".join((split_key[0].upper(), split_key[1].upper(), split_key[2], split_key[3]))
+
+    #         net.add_node(key, label=label, title=title, color=colour, value=2000)
+    #         net.add_edge(source=site, to=key)
+
+    #     return net.show("openghg_rankings.html")
+
+
+def rank_sources(site: str, species: str) -> RankSources:
+    """Retrieve datasources for a specific site and species. Returns a RankSources
+    object that can be used to set and modify ranking data.
+
+    Args:
+        site: Site code
+        species: Species
+    Returns:
+        RankSources: A RankSources object
+    """
+    cloud = running_in_cloud()
+    ranker = RankSources(cloud)
+    ranker.get_sources(site=site, species=species)
+
+    return ranker
